requirements.txt
pyxel==1.8.22
Pillow==9.1.0